Fix import structure
The "jopowa_vis" directory is a package containing the `app` module and
the `apps` package so importing this module or from this package either
has to happen via relative imports or by referencing the `jopowa_vis`
package. This commit opts for the latter solution.
Also, since "app.py" and "jopowa_vis.py" define modules and not
packages, they're not found by `find_packages` and have to be specified
directly via `py_modules`.