Fix documentation building
Some links didn't work because:

  a: they still contained "simnh" instead of "znes" and
  b: they contained an underscore instead of the hyphen in "jopowa-vis".

Also, since there is no PyPI package yet for `jopowa-vis`, the link
behind the badge for the package is dead so I had to comment out the
badge for now. Last but not least there's no "_static" folder and no
"modules.rst" file, so I had to remove the references to those from
"conf.py" and "index.rst" respectively.